refactor(__main__): move functions into separate modules (#135) (+linting)
